thesis_catalyze/requirements-pc.txt
Flask>=2.0
requests
python-dotenv
flask-cors